+++ sdk/storagediscovery/azure-mgmt-storagediscovery/azure/mgmt/storagediscovery/aio/operations/_operations.py
@@ -30,7 +30,7 @@ from azure.core.tracing.decorator_async import distributed_trace_async
 from azure.core.utils import case_insensitive_dict
 from azure.mgmt.core.exceptions import ARMErrorFormat
 
-from ... import models as _models
+from ... import models as _models, types as _types
 from ..._utils.model_base import SdkJSONEncoder, _deserialize, _failsafe_deserialize
 from ..._utils.serialization import Deserializer, Serializer
 from ...operations._operations import (
@@ -44,13 +44,12 @@ from ...operations._operations import (
 )
 from .._configuration import StorageDiscoveryMgmtClientConfiguration
 
-JSON = MutableMapping[str, Any]
 T = TypeVar("T")
 ClsType = Optional[Callable[[PipelineResponse[HttpRequest, AsyncHttpResponse], T, dict[str, Any]], Any]]
 List = list
 
 
-class StorageDiscoveryWorkspacesOperations:
+class StorageDiscoveryWorkspacesOperations:  # pylint: disable=docstring-missing-param
     """
     .. warning::
         **DO NOT** instantiate this class directly.
@@ -111,6 +110,7 @@ class StorageDiscoveryWorkspacesOperations:
         }
         _request.url = self._client.format_url(_request.url, **path_format_arguments)
 
+        _decompress = kwargs.pop("decompress", True)
         _stream = kwargs.pop("stream", False)
         pipeline_response: PipelineResponse = await self._client._pipeline.run(  # pylint: disable=protected-access
             _request, stream=_stream, **kwargs
@@ -125,11 +125,14 @@ class StorageDiscoveryWorkspacesOperations:
                 except (StreamConsumedError, StreamClosedError):
                     pass
             map_error(status_code=response.status_code, response=response, error_map=error_map)
-            error = _failsafe_deserialize(_models.ErrorResponse, response)
+            error = _failsafe_deserialize(
+                _models.ErrorResponse,
+                response,
+            )
             raise HttpResponseError(response=response, model=error, error_format=ARMErrorFormat)
 
         if _stream:
-            deserialized = response.iter_bytes()
+            deserialized = response.iter_bytes() if _decompress else response.iter_raw()
         else:
             deserialized = _deserialize(_models.StorageDiscoveryWorkspace, response.json())
 
@@ -171,7 +174,7 @@ class StorageDiscoveryWorkspacesOperations:
         self,
         resource_group_name: str,
         storage_discovery_workspace_name: str,
-        resource: JSON,
+        resource: _types.StorageDiscoveryWorkspace,
         *,
         content_type: str = "application/json",
         **kwargs: Any
@@ -184,7 +187,7 @@ class StorageDiscoveryWorkspacesOperations:
         :param storage_discovery_workspace_name: The name of the StorageDiscoveryWorkspace. Required.
         :type storage_discovery_workspace_name: str
         :param resource: Resource create parameters. Required.
-        :type resource: JSON
+        :type resource: ~azure.mgmt.storagediscovery.types.StorageDiscoveryWorkspace
         :keyword content_type: Body Parameter content-type. Content type parameter for JSON body.
          Default value is "application/json".
         :paramtype content_type: str
@@ -227,7 +230,7 @@ class StorageDiscoveryWorkspacesOperations:
         self,
         resource_group_name: str,
         storage_discovery_workspace_name: str,
-        resource: Union[_models.StorageDiscoveryWorkspace, JSON, IO[bytes]],
+        resource: Union[_models.StorageDiscoveryWorkspace, _types.StorageDiscoveryWorkspace, IO[bytes]],
         **kwargs: Any
     ) -> _models.StorageDiscoveryWorkspace:
         """Create a StorageDiscoveryWorkspace.
@@ -237,10 +240,10 @@ class StorageDiscoveryWorkspacesOperations:
         :type resource_group_name: str
         :param storage_discovery_workspace_name: The name of the StorageDiscoveryWorkspace. Required.
         :type storage_discovery_workspace_name: str
-        :param resource: Resource create parameters. Is one of the following types:
-         StorageDiscoveryWorkspace, JSON, IO[bytes] Required.
-        :type resource: ~azure.mgmt.storagediscovery.models.StorageDiscoveryWorkspace or JSON or
-         IO[bytes]
+        :param resource: Resource create parameters. Is either a StorageDiscoveryWorkspace type or a
+         IO[bytes] type. Required.
+        :type resource: ~azure.mgmt.storagediscovery.models.StorageDiscoveryWorkspace or
+         ~azure.mgmt.storagediscovery.types.StorageDiscoveryWorkspace or IO[bytes]
         :return: StorageDiscoveryWorkspace. The StorageDiscoveryWorkspace is compatible with
          MutableMapping
         :rtype: ~azure.mgmt.storagediscovery.models.StorageDiscoveryWorkspace
@@ -282,6 +285,7 @@ class StorageDiscoveryWorkspacesOperations:
         }
         _request.url = self._client.format_url(_request.url, **path_format_arguments)
 
+        _decompress = kwargs.pop("decompress", True)
         _stream = kwargs.pop("stream", False)
         pipeline_response: PipelineResponse = await self._client._pipeline.run(  # pylint: disable=protected-access
             _request, stream=_stream, **kwargs
@@ -296,11 +300,14 @@ class StorageDiscoveryWorkspacesOperations:
                 except (StreamConsumedError, StreamClosedError):
                     pass
             map_error(status_code=response.status_code, response=response, error_map=error_map)
-            error = _failsafe_deserialize(_models.ErrorResponse, response)
+            error = _failsafe_deserialize(
+                _models.ErrorResponse,
+                response,
+            )
             raise HttpResponseError(response=response, model=error, error_format=ARMErrorFormat)
 
         if _stream:
-            deserialized = response.iter_bytes()
+            deserialized = response.iter_bytes() if _decompress else response.iter_raw()
         else:
             deserialized = _deserialize(_models.StorageDiscoveryWorkspace, response.json())
 
@@ -342,7 +349,7 @@ class StorageDiscoveryWorkspacesOperations:
         self,
         resource_group_name: str,
         storage_discovery_workspace_name: str,
-        properties: JSON,
+        properties: _types.StorageDiscoveryWorkspaceUpdate,
         *,
         content_type: str = "application/json",
         **kwargs: Any
@@ -355,7 +362,7 @@ class StorageDiscoveryWorkspacesOperations:
         :param storage_discovery_workspace_name: The name of the StorageDiscoveryWorkspace. Required.
         :type storage_discovery_workspace_name: str
         :param properties: The resource properties to be updated. Required.
-        :type properties: JSON
+        :type properties: ~azure.mgmt.storagediscovery.types.StorageDiscoveryWorkspaceUpdate
         :keyword content_type: Body Parameter content-type. Content type parameter for JSON body.
          Default value is "application/json".
         :paramtype content_type: str
@@ -398,7 +405,7 @@ class StorageDiscoveryWorkspacesOperations:
         self,
         resource_group_name: str,
         storage_discovery_workspace_name: str,
-        properties: Union[_models.StorageDiscoveryWorkspaceUpdate, JSON, IO[bytes]],
+        properties: Union[_models.StorageDiscoveryWorkspaceUpdate, _types.StorageDiscoveryWorkspaceUpdate, IO[bytes]],
         **kwargs: Any
     ) -> _models.StorageDiscoveryWorkspace:
         """Update a StorageDiscoveryWorkspace.
@@ -408,10 +415,10 @@ class StorageDiscoveryWorkspacesOperations:
         :type resource_group_name: str
         :param storage_discovery_workspace_name: The name of the StorageDiscoveryWorkspace. Required.
         :type storage_discovery_workspace_name: str
-        :param properties: The resource properties to be updated. Is one of the following types:
-         StorageDiscoveryWorkspaceUpdate, JSON, IO[bytes] Required.
-        :type properties: ~azure.mgmt.storagediscovery.models.StorageDiscoveryWorkspaceUpdate or JSON
-         or IO[bytes]
+        :param properties: The resource properties to be updated. Is either a
+         StorageDiscoveryWorkspaceUpdate type or a IO[bytes] type. Required.
+        :type properties: ~azure.mgmt.storagediscovery.models.StorageDiscoveryWorkspaceUpdate or
+         ~azure.mgmt.storagediscovery.types.StorageDiscoveryWorkspaceUpdate or IO[bytes]
         :return: StorageDiscoveryWorkspace. The StorageDiscoveryWorkspace is compatible with
          MutableMapping
         :rtype: ~azure.mgmt.storagediscovery.models.StorageDiscoveryWorkspace
@@ -453,6 +460,7 @@ class StorageDiscoveryWorkspacesOperations:
         }
         _request.url = self._client.format_url(_request.url, **path_format_arguments)
 
+        _decompress = kwargs.pop("decompress", True)
         _stream = kwargs.pop("stream", False)
         pipeline_response: PipelineResponse = await self._client._pipeline.run(  # pylint: disable=protected-access
             _request, stream=_stream, **kwargs
@@ -467,11 +475,14 @@ class StorageDiscoveryWorkspacesOperations:
                 except (StreamConsumedError, StreamClosedError):
                     pass
             map_error(status_code=response.status_code, response=response, error_map=error_map)
-            error = _failsafe_deserialize(_models.ErrorResponse, response)
+            error = _failsafe_deserialize(
+                _models.ErrorResponse,
+                response,
+            )
             raise HttpResponseError(response=response, model=error, error_format=ARMErrorFormat)
 
         if _stream:
-            deserialized = response.iter_bytes()
+            deserialized = response.iter_bytes() if _decompress else response.iter_raw()
         else:
             deserialized = _deserialize(_models.StorageDiscoveryWorkspace, response.json())
 
@@ -528,7 +539,10 @@ class StorageDiscoveryWorkspacesOperations:
 
         if response.status_code not in [200, 204]:
             map_error(status_code=response.status_code, response=response, error_map=error_map)
-            error = _failsafe_deserialize(_models.ErrorResponse, response)
+            error = _failsafe_deserialize(
+                _models.ErrorResponse,
+                response,
+            )
             raise HttpResponseError(response=response, model=error, error_format=ARMErrorFormat)
 
         if cls:
@@ -589,7 +603,10 @@ class StorageDiscoveryWorkspacesOperations:
                 )
                 _next_request_params["api-version"] = self._config.api_version
                 _request = HttpRequest(
-                    "GET", urllib.parse.urljoin(next_link, _parsed_next_link.path), params=_next_request_params
+                    "GET",
+                    urllib.parse.urljoin(next_link, _parsed_next_link.path),
+                    headers=_headers,
+                    params=_next_request_params,
                 )
                 path_format_arguments = {
                     "endpoint": self._serialize.url(
@@ -602,7 +619,10 @@ class StorageDiscoveryWorkspacesOperations:
 
         async def extract_data(pipeline_response):
             deserialized = pipeline_response.http_response.json()
-            list_of_elem = _deserialize(List[_models.StorageDiscoveryWorkspace], deserialized.get("value", []))
+            list_of_elem = _deserialize(
+                List[_models.StorageDiscoveryWorkspace],
+                deserialized.get("value", []),
+            )
             if cls:
                 list_of_elem = cls(list_of_elem)  # type: ignore
             return deserialized.get("nextLink") or None, AsyncList(list_of_elem)
@@ -618,7 +638,10 @@ class StorageDiscoveryWorkspacesOperations:
 
             if response.status_code not in [200]:
                 map_error(status_code=response.status_code, response=response, error_map=error_map)
-                error = _failsafe_deserialize(_models.ErrorResponse, response)
+                error = _failsafe_deserialize(
+                    _models.ErrorResponse,
+                    response,
+                )
                 raise HttpResponseError(response=response, model=error, error_format=ARMErrorFormat)
 
             return pipeline_response
@@ -674,7 +697,10 @@ class StorageDiscoveryWorkspacesOperations:
                 )
                 _next_request_params["api-version"] = self._config.api_version
                 _request = HttpRequest(
-                    "GET", urllib.parse.urljoin(next_link, _parsed_next_link.path), params=_next_request_params
+                    "GET",
+                    urllib.parse.urljoin(next_link, _parsed_next_link.path),
+                    headers=_headers,
+                    params=_next_request_params,
                 )
                 path_format_arguments = {
                     "endpoint": self._serialize.url(
@@ -687,7 +713,10 @@ class StorageDiscoveryWorkspacesOperations:
 
         async def extract_data(pipeline_response):
             deserialized = pipeline_response.http_response.json()
-            list_of_elem = _deserialize(List[_models.StorageDiscoveryWorkspace], deserialized.get("value", []))
+            list_of_elem = _deserialize(
+                List[_models.StorageDiscoveryWorkspace],
+                deserialized.get("value", []),
+            )
             if cls:
                 list_of_elem = cls(list_of_elem)  # type: ignore
             return deserialized.get("nextLink") or None, AsyncList(list_of_elem)
@@ -703,7 +732,10 @@ class StorageDiscoveryWorkspacesOperations:
 
             if response.status_code not in [200]:
                 map_error(status_code=response.status_code, response=response, error_map=error_map)
-                error = _failsafe_deserialize(_models.ErrorResponse, response)
+                error = _failsafe_deserialize(
+                    _models.ErrorResponse,
+                    response,
+                )
                 raise HttpResponseError(response=response, model=error, error_format=ARMErrorFormat)
 
             return pipeline_response
@@ -711,7 +743,7 @@ class StorageDiscoveryWorkspacesOperations:
         return AsyncItemPaged(get_next, extract_data)
 
 
-class Operations:
+class Operations:  # pylint: disable=docstring-missing-param
     """
     .. warning::
         **DO NOT** instantiate this class directly.
@@ -777,7 +809,10 @@ class Operations:
                 )
                 _next_request_params["api-version"] = self._config.api_version
                 _request = HttpRequest(
-                    "GET", urllib.parse.urljoin(next_link, _parsed_next_link.path), params=_next_request_params
+                    "GET",
+                    urllib.parse.urljoin(next_link, _parsed_next_link.path),
+                    headers=_headers,
+                    params=_next_request_params,
                 )
                 path_format_arguments = {
                     "endpoint": self._serialize.url(
@@ -790,7 +825,10 @@ class Operations:
 
         async def extract_data(pipeline_response):
             deserialized = pipeline_response.http_response.json()
-            list_of_elem = _deserialize(List[_models.Operation], deserialized.get("value", []))
+            list_of_elem = _deserialize(
+                List[_models.Operation],
+                deserialized.get("value", []),
+            )
             if cls:
                 list_of_elem = cls(list_of_elem)  # type: ignore
             return deserialized.get("nextLink") or None, AsyncList(list_of_elem)
@@ -806,7 +844,10 @@ class Operations:
 
             if response.status_code not in [200]:
                 map_error(status_code=response.status_code, response=response, error_map=error_map)
-                error = _failsafe_deserialize(_models.ErrorResponse, response)
+                error = _failsafe_deserialize(
+                    _models.ErrorResponse,
+                    response,
+                )
                 raise HttpResponseError(response=response, model=error, error_format=ARMErrorFormat)
 
             return pipeline_response