+++ sdk/storagediscovery/azure-mgmt-storagediscovery/generated_tests/test_storage_discovery_mgmt_storage_discovery_workspaces_operations.py
@@ -44,6 +44,14 @@ class TestStorageDiscoveryMgmtStorageDiscoveryWorkspacesOperations(AzureMgmtReco
                         {"displayName": "str", "resourceTypes": ["str"], "tagKeysOnly": ["str"], "tags": {"str": "str"}}
                     ],
                     "workspaceRoots": ["str"],
+                    "capabilities": {
+                        "azureBlobStorage": {
+                            "capacityDetails": {"status": "str"},
+                            "prefixConfigurations": [
+                                {"containerName": "str", "storageAccountName": "str", "prefix": "str"}
+                            ],
+                        }
+                    },
                     "description": "str",
                     "provisioningState": "str",
                     "sku": "str",
@@ -72,6 +80,14 @@ class TestStorageDiscoveryMgmtStorageDiscoveryWorkspacesOperations(AzureMgmtReco
             storage_discovery_workspace_name="str",
             properties={
                 "properties": {
+                    "capabilities": {
+                        "azureBlobStorage": {
+                            "capacityDetails": {"status": "str"},
+                            "prefixConfigurations": [
+                                {"containerName": "str", "prefix": "str", "storageAccountName": "str"}
+                            ],
+                        }
+                    },
                     "description": "str",
                     "scopes": [
                         {"displayName": "str", "resourceTypes": ["str"], "tagKeysOnly": ["str"], "tags": {"str": "str"}}